\name{get_option_parser_evaluation}
\alias{get_option_parser_evaluation}
\title{Internal - parsing options for REPTILE_evaluate_prediction.R
}
\description{
  Internal function used to parsing options for
  "REPTILE_evaluate_prediction.R" script in the REPTILE enhancer
  prediction pipeline:

  \url{https://github.com/yupenghe/REPTILE}
}
\usage{
get_option_parser_evaluation()
}
\value{
  An instance of the OptionParser class.
}
\author{
  Yupeng He \email{yupeng.he.bioinfo@gmail.com}
}
\keyword{ option parser }

