\name{get_option_parser_training}
\alias{get_option_parser_training}
\title{
  Internal - parsing options for REPTILE_train.R
}
\description{
  Internal function used to parsing options
  for "REPTILE_train.R" script in the REPTILE enhancer
  prediction pipeline:

  \url{https://github.com/yupenghe/REPTILE}
}
\usage{
get_option_parser_training()
}
\value{
  An instance of the OptionParser class.
}
\author{
  Yupeng He \email{yupeng.he.bioinfo@gmail.com}
}
\keyword{ option parser }

